Renamed conformation_processing.rs to log_transaction_service.rs
